Fix narrowing_error on 64-bit SUB with an INT_MIN immediate (#1229)
`r0 -= -0x80000000` aborted the analysis with `error: narrowing_error`:

    0:  b7 00 00 00 00 00 00 00 r0 = 0x0
    1:  17 00 00 00 00 00 00 80 r0 -= -0x80000000
    2:  95 00 00 00 00 00 00 00 exit

`EbpfTransformer::add` took the immediate as an `int`, and the SUB case reached
it through `add(bin.dst, gsl::narrow<int>(-imm), ...)`. `imm` is an `int64_t`
holding a sign-extended 32-bit immediate, so for `imm == INT32_MIN` the negation
is 2147483648, one past `INT_MAX`, and `gsl::narrow` threw. The value is
otherwise ordinary: nothing downstream of `add` needs a 32-bit parameter, since
every use converts to `Number`.

Widen the parameter to `int64_t` and negate there. `|imm| <= 2^31` for both the
64-bit path (sign-extended imm32) and the 32-bit path (`narrow_cast<int32_t>`),
so the negation cannot overflow.

Verified against the reporter's object file, which now passes and leaves
r0.svalue = 2147483648. The two new subtract.yaml cases fail with
`Exception: narrowing_error` when the source fix is reverted; the add.yaml case
pins the ADD side of the same boundary, which was already correct.

Reported-by: fuzzing, https://github.com/vbpf/prevail/issues/1229

+++ src/crab/ebpf_transformer.cpp
@@ -105,7 +105,7 @@ class EbpfTransformer final {
     void do_mem_store(const Mem& b, const LinearExpression& val_svalue, const LinearExpression& val_uvalue,
                       const std::optional<Reg>& opt_val_reg);
 
-    void add(const Reg& dst_reg, int imm, int finite_width);
+    void add(const Reg& dst_reg, int64_t imm, int finite_width);
 
     void shl(const Reg& dst_reg, int imm, int finite_width);
 
@@ -1129,7 +1129,7 @@ void EbpfTransformer::recompute_stack_numeric_size(TypeToNumDomain& state, const
     recompute_stack_numeric_size(state, reg_type(reg));
 }
 
-void EbpfTransformer::add(const Reg& dst_reg, const int imm, const int finite_width) {
+void EbpfTransformer::add(const Reg& dst_reg, const int64_t imm, const int finite_width) {
     const auto dst = reg_pack(dst_reg);
     if (dom.state.may_have_type(dst_reg, T_NUM)) {
         dom.state.values->add_overflow(dst.svalue, dst.uvalue, imm, finite_width);
@@ -1258,13 +1258,15 @@ void EbpfTransformer::operator()(const Bin& bin) {
             if (imm == 0) {
                 return;
             }
-            add(bin.dst, gsl::narrow<int>(imm), finite_width);
+            add(bin.dst, imm, finite_width);
             break;
         case Bin::Op::SUB:
             if (imm == 0) {
                 return;
             }
-            add(bin.dst, gsl::narrow<int>(-imm), finite_width);
+            // Negate in 64 bits: imm is a sign-extended 32-bit value, so -INT32_MIN
+            // is representable here even though it does not fit in an int.
+            add(bin.dst, -imm, finite_width);
             break;
         case Bin::Op::MUL:
             dom.state.values->mul(dst.svalue, dst.uvalue, imm, finite_width);
